requirements-dev.txt
appdirs==1.4.4
distlib==0.3.1
filelock==3.0.12
humanfriendly==8.2
importlib-metadata==1.7.0
numpy==1.19.0
pandas==1.1.2
psycopg2==2.8.6
py4j==0.10.9
pyarrow==0.17.1
pyspark==3.0.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
SQLAlchemy==1.3.19
virtualenv==20.0.26
zipp==3.1.0
